@@ -16,7 +16,7 @@
 
 package com.stfalcon.chatkit.commons;
 
-import android.support.v7.widget.RecyclerView;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.View;
 
 /**
