@@ -1,7 +1,7 @@
 package com.stfalcon.chatkit.sample.utils;
 
 import android.content.Context;
-import android.support.annotation.StringRes;
+import androidx.annotation.StringRes;
 import android.widget.Toast;
 
 /*
